chore: Refactored RowSelection into folder with separated base, types, and utils
